+++ components/hn-table/hn-table.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 
-import { NzTableModule } from 'ng-zorro-antd';
+import { NzTableModule } from 'ng-zorro-antd/table';
 import { HnTableComponent } from './hn-table.component';
 
 @NgModule({