+++ src/app/pages/demo/table/table-render/table-render.component.ts
@@ -7,15 +7,6 @@ import { columns, item } from '../columns';
 export class TableRenderComponent implements OnInit {
   @ViewChild('table', { static: true }) table: any;
 
-  selections: any[] = [
-    {
-      text: 'half',
-      onSelect: () => {
-        this.handleSelect();
-      }
-    }
-  ];
-
   filters: any[] = [
     {
       text: '男',
@@ -28,6 +19,11 @@ export class TableRenderComponent implements OnInit {
   ];
 
   columns = [
+    {
+      showCheckbox: true,
+      width: '60px',
+      left: '0px'
+    },
     { title: '名称', key: 'name' },
     {
       title: '性别',
@@ -39,8 +35,26 @@ export class TableRenderComponent implements OnInit {
         this.handleFilter(event, 'gender');
       }
     },
-    { title: '年龄', key: 'age', showSort: true },
-    { title: '出生日期', key: 'createTime', showSort: true },
+    {
+      title: '年龄',
+      key: 'age',
+      showSort: true,
+      sortOrderChange: (event: any) => {
+        const data = { key: 'age', value: event };
+        this.handleSortChange(data);
+      }
+    },
+    { title: '身高', key: 'height' },
+    { title: '身份', key: 'identity' },
+    {
+      title: '出生日期',
+      key: 'createTime',
+      showSort: true,
+      sortOrderChange: (event: any) => {
+        const data = { key: 'createTime', value: event };
+        this.handleSortChange(data);
+      }
+    },
     { title: '链接', key: 'link', renderKey: 'linkTemp', thRenderKey: 'linkThTemp' }
   ];
   tableData: any[] = [];
@@ -65,6 +79,8 @@ export class TableRenderComponent implements OnInit {
         ...item,
         name: `${item.name}${i}`,
         age: Number((Math.random() * 100).toFixed(0)),
+        height: '7尺7寸',
+        identity: '季汉昭烈帝',
         createTime: Number((Math.random() * 1000).toFixed(0)),
         gender: i % 2 === 1 ? '男' : '女'
       });
@@ -83,13 +99,6 @@ export class TableRenderComponent implements OnInit {
    */
   handleCheckChange(event: any): void {
     console.log('选中', event);
-  }
-
-  /**
-   * 选择
-   */
-  handleSelect(): void {
-    console.log('选中');
   }
 
   /**